Implement New Favorites List Command
It's possible the favorites list may be very long. In the event this
occurs, care must be taken when displaying the information as Discord
imposes limits on the size of embeds (and more generally, messages).

This implementation adds a library which makes paginating embeds easy so
that the output of this command can handle an arbitrary amount of
favorites. Furthermore, it switches the output to ephemeral to help
prevent clutter in channels.

Signed-off-by: Cory J Reid <11530109+coryjreid@users.noreply.github.com>